% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StataExport.R
\name{export_dynamic_crosstab}
\alias{export_dynamic_crosstab}
\title{Epoeting crosstab of stata datasets into excel}
\usage{
export_dynamic_crosstab(file_path, variables, output_excel, sheet_name)
}
\arguments{
\item{sheet_name}{}
}
\description{
Epoeting crosstab of stata datasets into excel
}
